+++ site/docs/pages/token/token-search.mdx
@@ -1,6 +1,5 @@
 import { TokenSearch } from '@coinbase/onchainkit/token';
 import App from '../App';
-import { tokens } from './tokens';
 
 # `<TokenSearch />`
